Fix xml parsing error
Previous code was attempting to parse html to xml but it's more prone to parsing error.  The InnerHtml property of a Form object returns input elements without ending slash which isn't supported by the XML parser and the quick workaround of replacing "> by "/> wasn't working because the attribute values weren't surrounded by quotes.